+++ yamlator/validators/any_type_validator.py
@@ -1,6 +1,5 @@
 """Validator for handling the any data type"""
 
-
 from yamlator.types import Data
 from yamlator.types import RuleType
 from yamlator.types import SchemaTypes
@@ -16,11 +15,11 @@ class AnyTypeValidator(Validator):
         """Validate any rules that have the data marked as the `any` type
 
         Args:
-            key (str): The key to the data
-            data (Data): The data to validate
+            key (str): The data field name
+            data (yamlator.types.Data): The data to validate
             parent (str): The parent key of the data
-            rtype (RuleType): The type assigned to the rule that will be
-                applied to the data
+            rtype (yamlator.types.RuleType): The type assigned to the
+                rule that will be applied to the data
             is_required (bool, optional): Indicates if the rule is required
         """
         is_any_type = (rtype.schema_type == SchemaTypes.ANY)