+++ yamlator/validators/builtin_type_validator.py
@@ -19,8 +19,8 @@ class BuiltInTypeValidator(Validator):
         """BuiltInTypeValidator init
 
         Args:
-            violations (deque): Contains violations that have been detected
-                whilst processing the data
+            violations (collections.deque): Contains violations that have
+                been detected whilst processing the data
         """
         super().__init__(violations)
         self._built_in_lookups = {
@@ -39,11 +39,11 @@ class BuiltInTypeValidator(Validator):
         to the list of violations
 
         Args:
-            key (str): The key to the data
-            data (Data): The data to validate
+            key (str): The data field name
+            data (yamlator.types.Data): The data to validate
             parent (str): The parent key of the data
-            rtype (RuleType): The type assigned to the rule that will be
-                applied to the data
+            rtype (yamlator.types.RuleType): The type assigned to the
+                rule that will be applied to the data
             is_required (bool, optional): Indicates if the rule is required
         """
         buildin_type = self._built_in_lookups.get(rtype.schema_type)