+++ yamlator/validators/entry_point_validator.py
@@ -9,6 +9,7 @@ from yamlator.types import Rule
 from yamlator.types import RuleType
 from yamlator.types import YamlatorRuleset
 from yamlator.violations import StrictEntryPointViolation
+from yamlator.utils import is_keyless_rule
 from .base_validator import Validator
 
 
@@ -19,11 +20,14 @@ class EntryPointValidator(Validator):
         """EntryPointValidator init
 
         Args:
-            violations (deque): contains violations that have been
+            violations (collections.deque): Contains violations that have been
                 detected whilst processing the data
-            entry_point (YamlatorRuleset): The entry point ruleset
+
+            entry_point (yamlator.types.YamlatorRuleset): The entry
+                point ruleset
         """
         self._entry_point = entry_point
+        self._flat_validator = None
         super().__init__(violations)
 
     def validate(self, key: str, data: Data, parent: str, rtype: RuleType,
@@ -32,18 +36,26 @@ class EntryPointValidator(Validator):
         rule by calling the next validator in the chain
 
         Args:
-            key (str): The key to the data
-            data (Data): The data to validate
+            key (str): The data field name
+            data (yamlator.types.Data): The data to validate
             parent (str): The parent key of the data
-            rtype (RuleType): The type assigned to the rule that will be
-                applied to the data
+            rtype (yamlator.types.RuleType): The type assigned to the
+                rule that will be applied to the data
             is_required (bool, optional): Indicates if the rule is required
         """
+        # These are not used by the `EntryPointValidator`
         del key
         del rtype
         del is_required
 
         rules = self._entry_point.rules
+        if not rules:
+            return
+
+        has_validated = self._validate_keyless_data(data, parent, rules)
+        if has_validated:
+            return
+
         if self._entry_point.is_strict:
             self._handle_strict_mode(data, rules)
 
@@ -52,6 +64,24 @@ class EntryPointValidator(Validator):
 
             super().validate(rule.name, sub_data, parent,
                              rule.rtype, rule.is_required)
+
+    def _validate_keyless_data(self, data: Data,
+                               parent: str,
+                               rules: Iterable[Rule]):
+        # If there is more than 1 rule, that we are not
+        # dealing with a keyless root
+        if len(rules) > 1:
+            return False
+
+        rule: Rule = rules[0]
+        if not is_keyless_rule(rule):
+            return False
+
+        # Run the validation here instead since we are dealing
+        # with an object that does not have a root key. E.g a list
+        super().validate(rule.name, data, parent,
+                         rule.rtype, rule.is_required)
+        return True
 
     def _handle_strict_mode(self, data: dict, rules: Iterable[Rule]):
         rule_fields = {rule.name for rule in rules}